backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
qdrant-client
psycopg2-binary
python-dotenv
openai
tenacity
google-generativeai
openai-agents
cohere
requests
beautifulsoup4
